+++ GUI/LightingGUI/arraycolorspage.h
@@ -4,9 +4,13 @@
 #define ARRAYCOLORSPAGE_H
 
 #include "lightingpage.h"
+#include "icondata.h"
+
 #include <QWidget>
 #include <QDebug>
 #include <QToolButton>
+#include <QPushButton>
+#include <QGridLayout>
 
 namespace Ui {
 class ArrayColorsPage;
@@ -16,14 +20,15 @@ class ArrayColorsPage;
  * \copyright
  * Copyright (C) 2015 - 2016. All Rights MIT Licensed.
  *
- * \brief The ArrayColorsPage class uses the color array from the arduino's RoutinesRGB library.
+ * \brief The ArrayColorsPage class uses the color array routines from the arduino's RoutinesRGB library.
  *
- * The bar at the top of the page determines how many of the colors from the array to use. For example,
- * when its set to `2` it will use only the first `2` colors. Setting it to `5` will use `5`
- * array colors, etc.
+ * The page contains an interface to change the individual array colors and a series of buttons
+ * that change the mode.
  *
- * \todo Add the ability to set each array color individually.
- *
+ * The top interface is split into three parts: a ColorPicker, a color array preview, and
+ * a LightsSlider. The ColorPicker chooses the color that is currently selected from the color
+ * array preview. The LightsSlider below the preview is used to choose how many colors are active
+ * while using an array mode.
  */
 class ArrayColorsPage : public QWidget, public LightingPage
 {
@@ -32,8 +37,8 @@ class ArrayColorsPage : public QWidget, public LightingPage
 public:
     explicit ArrayColorsPage(QWidget *parent = 0);
     ~ArrayColorsPage();
-    void highlightButton(ELightingMode lightingMode);
 
+    void highlightModeButton(ELightingMode lightingMode);
 
 signals:
     /*!
@@ -48,7 +53,9 @@ signals:
  */
 public slots:
     void modeChanged(int);
-    void colorCountChanged(int);
+    void colorsUsedChanged(int);
+    void colorChanged(QColor);
+    void selectArrayColor(int);
 
 protected:
     void showEvent(QShowEvent *event);
@@ -56,12 +63,25 @@ protected:
 private:
     Ui::ArrayColorsPage *ui;
 
+    std::shared_ptr<std::vector<QPushButton *> > mArrayColorsButtons;
+    std::shared_ptr<std::vector<IconData> > mArrayColorsIconData;
+
+    /*!
+     * \brief mGreyIcon used for greying out icons
+     */
+    IconData mGreyIcon;
+
     /*!
      * \brief mPageButtons pointers to all the main buttons, used
      * to iterate through them quickly.
      */
-    std::shared_ptr<std::vector<QToolButton*> > mPageButtons;
+    std::shared_ptr<std::vector<QToolButton *> > mPageButtons;
 
+    /*!
+     * \brief mCurrentColorPickerIndex current index being set by the GUI.
+     */
+    int mCurrentColorPickerIndex;
+    void updateColorArray();
     void updateIcons();
 };