+++ GUI/LightingGUI/singlecolorpage.cpp
@@ -15,10 +15,6 @@ SingleColorPage::SingleColorPage(QWidget *parent) :
     ui(new Ui::SingleColorPage) {
     ui->setupUi(this);
 
-
-    QPixmap pixmap = QPixmap(":/images/color_wheel.png");
-    ui->colorPicker->setPixmap(pixmap.scaled(175, 175, Qt::KeepAspectRatio));
-
     mSolidData = IconData(ui->solidButton->iconSize().width(), ui->solidButton->iconSize().height());
     mFadeData = IconData(ui->fadeButton->iconSize().width(), ui->fadeButton->iconSize().height());
     mGlimmerData = IconData(ui->glimmerButton->iconSize().width(), ui->glimmerButton->iconSize().height());
@@ -45,23 +41,8 @@ SingleColorPage::SingleColorPage(QWidget *parent) :
 
     connect(signalMapper, SIGNAL(mapped(int)), this, SLOT(modeChanged(int)));
 
-    // set up the sliders
-    ui->sliderR->setSliderColorBackground({255, 0, 0});
-    ui->sliderG->setSliderColorBackground({0, 255, 0});
-    ui->sliderB->setSliderColorBackground({0, 0, 255});
-
-    ui->sliderR->slider->setRange(0, 255);
-    ui->sliderG->slider->setRange(0, 255);
-    ui->sliderB->slider->setRange(0, 255);
-
-    connect(ui->colorPicker, SIGNAL(colorWheelUpdate(QColor)),
-            this, SLOT(colorWheelChanged(QColor)));
-    connect(ui->sliderR, SIGNAL(valueChanged(int)),
-            this, SLOT(rChanged(int)));
-    connect(ui->sliderG, SIGNAL(valueChanged(int)),
-            this, SLOT(gChanged(int)));
-    connect(ui->sliderB, SIGNAL(valueChanged(int)),
-            this, SLOT(bChanged(int)));
+    ui->colorPicker->chooseLayout(ELayoutColorPicker::eFullLayout);
+    connect(ui->colorPicker, SIGNAL(colorUpdate(QColor)), this, SLOT(colorChanged(QColor)));
 }
 
 
@@ -69,15 +50,11 @@ SingleColorPage::~SingleColorPage() {
     delete ui;
 }
 
-
-void SingleColorPage::updateColorPreview() {
-    updateIcons();
-    // emit the updates to other UI elements
-    emit updateMainIcons();
+void SingleColorPage::chooseColor(QColor color) {
+  ui->colorPicker->chooseColor(color);
 }
 
-
-void SingleColorPage::highlightButton(ELightingMode lightingMode) {
+void SingleColorPage::highlightModeButton(ELightingMode lightingMode) {
     for (uint i = 0; i < mPageButtons->size(); i++) {
         QToolButton* button =(*mPageButtons.get())[i];
         button->setChecked(false);
@@ -98,8 +75,17 @@ void SingleColorPage::highlightButton(ELightingMode lightingMode) {
 // Slots
 // ----------------------------
 
-void SingleColorPage::setNewColor() {
-    mComm->sendMainColorChange(mData->color().red(), mData->color().green(), mData->color().blue());
+void SingleColorPage::modeChanged(int newMode) {
+    mCurrentMode = (ELightingMode)newMode;
+    mData->currentMode(mCurrentMode);
+    mComm->sendModeChange(mCurrentMode);
+    highlightModeButton(mCurrentMode);
+}
+
+void SingleColorPage::colorChanged(QColor color) {
+    mData->color(color);
+
+    mComm->sendMainColorChange(mData->color());
     if (!(mData->currentMode() == ELightingMode::eLightingModeSingleBlink
             || mData->currentMode() == ELightingMode::eLightingModeSingleConstant
             || mData->currentMode() == ELightingMode::eLightingModeSingleFade
@@ -107,76 +93,6 @@ void SingleColorPage::setNewColor() {
         mData->currentMode(ELightingMode::eLightingModeSingleGlimmer);
         mComm->sendModeChange(ELightingMode::eLightingModeSingleGlimmer);
     }
-}
-
-void SingleColorPage::showEvent(QShowEvent *) {
-  mCurrentMode = mData->currentMode();
-  highlightButton(mCurrentMode);
-}
-
-
-void SingleColorPage::modeChanged(int newMode) {
-    mCurrentMode = (ELightingMode)newMode;
-    mData->currentMode(mCurrentMode);
-    mComm->sendModeChange(mCurrentMode);
-    highlightButton(mCurrentMode);
-}
-
-
-void SingleColorPage::colorWheelChanged(QColor color) {
-    mData->color(color);
-    updateColorPreview();
-    setNewColor();
-}
-
-void SingleColorPage::rChanged(int newR) {
-    mData->color(QColor(newR, mData->color().green(), mData->color().blue()));
-    updateColorPreview();
-    setNewColor();
-}
-
-void SingleColorPage::gChanged(int newG) {
-    mData->color(QColor(mData->color().red(), newG, mData->color().blue()));
-    updateColorPreview();
-    setNewColor();
-}
-
-void SingleColorPage::bChanged(int newB) {
-    mData->color(QColor(mData->color().red(), mData->color().green(), newB));
-    updateColorPreview();
-    setNewColor();
-}
-
-
-// ----------------------------
-// Protected
-// ----------------------------
-
-
-void SingleColorPage::resizeEvent(QResizeEvent *event) {
-    Q_UNUSED(event);
-    QPixmap pixmap = QPixmap(":/images/color_wheel.png");
-    ui->colorPicker->setPixmap(pixmap.scaled(ui->colorPicker->height(), ui->colorPicker->height(), Qt::KeepAspectRatio, Qt::SmoothTransformation));
-}
-
-// ----------------------------
-// Private
-// ----------------------------
-
-void SingleColorPage::updateIcons() {
-    QColor color(mData->color());
-
-    ui->sliderR->slider->blockSignals(true);
-    ui->sliderR->slider->setValue(color.red());
-    ui->sliderR->slider->blockSignals(false);
-
-    ui->sliderG->slider->blockSignals(true);
-    ui->sliderG->slider->setValue(color.green());
-    ui->sliderG->slider->blockSignals(false);
-
-    ui->sliderB->slider->blockSignals(true);
-    ui->sliderB->slider->setValue(color.blue());
-    ui->sliderB->slider->blockSignals(true);
 
     mSolidData.setSolidColor(color);
     ui->solidButton->setIcon(mSolidData.renderAsQPixmap());
@@ -192,5 +108,17 @@ void SingleColorPage::updateIcons() {
     mBlinkData.setSolidColor(color);
     mBlinkData.addBlink();
     ui->blinkButton->setIcon(mBlinkData.renderAsQPixmap());
+
+    emit updateMainIcons();
+}
+
+// ----------------------------
+// Protected
+// ----------------------------
+
+
+void SingleColorPage::showEvent(QShowEvent *) {
+  mCurrentMode = mData->currentMode();
+  highlightModeButton(mCurrentMode);
 }